Fixed bug and improved path creation.
Made the path creation easier with the possibility to use the same spawnPoints for different paths.
Optimized the way the npc gets the point to walk to on path. They now choose one at random each time the should visit a new pathpoint (affects guards the most)

Fixed the bug with some civilians still walking while being sucked (was a missed transition in the state machine for them)